+++ documentation/contributing/changelist.md
@@ -8,3 +8,4 @@
 * 0.8.0  - Variable Declarations, reassignments and strong typing.  Readonly variables
 * 0.9.0  - Unnesting subshells, makefile integration and Brew Formula integration
 * 0.10.0 - Automatic strict mode
+* 0.11.0 - Added integer arithmetic, --verbose option
