+++ gradle.properties
@@ -2,7 +2,7 @@
 kotlin.code.style=official
 
 # Version properties
-version.bashpile=0.10.0
+version.bashpile=0.11.0
 # gradle version specified at gradle/wrapper/gradle-wrapper.properties
 version.antlr=4.13.2
 version.clikt=5.0.1
