+++ src/main/kotlin/org/bashpile/core/antlr/package-info.md
@@ -1,0 +1,1 @@
+For interfacing with Antlr4.
